Use helper function for intake log pull
This involves a fair number of changes (and a couple of blocks moved for clarity), but is really just as it says in the commit.
Right now your participant list is being derived from the sample intake log, but you could pull it from the paris & umbrella trackers and use it to filter your initial intake log query.
For now, I haven't made any changes like that. I was able to remove a lot of the filtering and the result cleaning since that happens in the query_intake function. You'll also see a couple of column name changes. Possible that you were not using this because it initially dropped the sample sharing columns, but don't hesitate to make small edits to helpers.py if they simplify your code (as they do in this case).