Pvi dashboarding (#59)
* Create pvi_dashboard.py

* start pvi_dashboard.py

* join cohorts within each study

i don't know if concat is the right function to use in line 118 but i don't think it should be join there too?

* create output with pvi_dashboard.py

it runs now! much reshaping needed

* added a sars-cov-2 specific tab to pvi_dashboard

this tab will be great for immune history figures! so far just apollo and paris, more to come!

* clean up covid tab in pvi dashboard

* output pvi dashboard to cross project

* add lookup tab to pvi dashboard

* Refactor pvi_dashboard

* add make_sheet function to pvi_dashboard

can add a sheet containing a given list of columns, in progress

* remove testing for make_sheet

---------

Co-authored-by: Charles Gleason <36055314+crgl@users.noreply.github.com>